Fix the two InputParserState/InputMode interop sites T4 missed
Five compile errors, from a survey that checked for integer interop by variable
name rather than by shape. Both real:

Four drivers index a static error-string table with the parser state —
errors[state]. That is static_cast<size_t> now.

LangTable::specific_key_exists tested `_mood` for truthiness, which meant "not
INPUT_MODE_NONE" because that enumerator is 0. It says so now.

A REAL DEFECT CAME OUT OF THE FIRST ONE, and it is preserved rather than fixed.
The error tables hold NINE strings; InputParserState has TEN enumerators. So
errors[STATE_DONE] reads one past the end of the array, and every message from
STATE_WANT_OPTIONAL onward is off by one — STATE_WANT_OPTIONAL returns
BAD_EXTRA's text, OVERSTEP returns CONJ_ERROR's, and so on. All four call sites
carry the explanation. Fixing it changes what the parser reports, so it is a
decision rather than a rider on a keyword change.

The lesson for the remaining enum work: grep for the SHAPE of the interop —
subscripting, boolean context, arithmetic — not for the names of the variables
you expect to find it in.

+++ NeuronClient/InputDriverAlias.cpp
@@ -55,7 +55,14 @@ static string errors[] = {"An unknown error occurred.",
                           "",
                           "There was no parsing error."};
 
-const string& AliasInputDriver::getLastParseError(InputParserState state) { return errors[state]; }
+// PRE-EXISTING AND PRESERVED: this table has NINE strings and
+// InputParserState has TEN enumerators, so errors[STATE_DONE] reads one past
+// the end, and every message from STATE_WANT_OPTIONAL on is off by one —
+// STATE_WANT_OPTIONAL gets BAD_EXTRA's text, and so on. The cast below makes
+// the indexing explicit; it does not change which string comes back. Surfaced
+// by language-hygiene T4, which had to look at the indexing to compile it.
+// Fixing it is a behaviour change and is recorded on that task, not done here.
+const string& AliasInputDriver::getLastParseError(InputParserState state) { return errors[static_cast<size_t>(state)]; }
 
 
 bool AliasInputDriver::getInputDescription(InputSpec const& spec, InputDescription& desc)
