+++ NeuronClient/InputDriverSimple.cpp
@@ -103,4 +103,11 @@ InputParserState SimpleInputDriver::parseExtraToken(std::string const& token, In
   return InputParserState::STATE_BAD_EXTRA; // Got into InputParserState::STATE_WANT_MODIFIER with no handler.
 }
 
-const string& SimpleInputDriver::getLastParseError(InputParserState state) { return errors[state]; }
+// PRE-EXISTING AND PRESERVED: this table has NINE strings and
+// InputParserState has TEN enumerators, so errors[STATE_DONE] reads one past
+// the end, and every message from STATE_WANT_OPTIONAL on is off by one —
+// STATE_WANT_OPTIONAL gets BAD_EXTRA's text, and so on. The cast below makes
+// the indexing explicit; it does not change which string comes back. Surfaced
+// by language-hygiene T4, which had to look at the indexing to compile it.
+// Fixing it is a behaviour change and is recorded on that task, not done here.
+const string& SimpleInputDriver::getLastParseError(InputParserState state) { return errors[static_cast<size_t>(state)]; }
